# Speed up build with WIN32_LEAN_AND_MEAN