This disables including rarely used headers.

This takes a full re-build of the main project from:
```
3>     1001 ms  Link                                       1 calls
3>     7387 ms  CL                                         1 calls
```
to:
```
3>     1035 ms  Link                                       1 calls
3>     6466 ms  CL                                         1 calls
```